Move streaming tests into separate module
